refactor(rfence): remove unnecessary synchronization in rfence extension implementation
